Add color and variables to monitoring tab
