+++ Localize/lang/strings.json
@@ -20,6 +20,7 @@
   "/VcZ9g": "This connector has multiple versions, built-in and Azure-hosted. Use built-in for the best performance, connection-string authentication, and other features. Use Azure-hosted for other authentication options.",
   "/WW7If": "True",
   "/doURb": "Convert the input to an array",
+  "/n13VL": "Properties",
   "/qchXQ": "Required. The collection to join items from.",
   "/qu3zt": "0",
   "/rjP4U": "(UTC) Coordinated Universal Time",
@@ -43,6 +44,7 @@
   "0hxr5j": "(UTC-05:00) Chetumal",
   "0i/6TR": "Required. The string to convert to upper casing. If a character in the string does not have an uppercase equivalent, the character is included unchanged in the returned string.",
   "0meqK1": "(UTC-05:00) Bogota, Lima, Quito, Rio Branco",
+  "0oebOm": "Outputs",
   "0p+pJq": "Returns the remainder after dividing the two numbers (modulo)",
   "0qV0Qe": "Required. The string that may contain the value.",
   "0qcKzG": "{parameterName} is no longer present in the operation schema. It should be removed before the workflow is re-saved.",
@@ -449,6 +451,7 @@
   "PBvR9Q": "(UTC-03:00) Araguaina",
   "PF87Ew": "Succeeded",
   "PKMkb/": "Required. The number to add to Summand 1.",
+  "PORNMZ": "Inputs",
   "PP63jY": "more panels",
   "PQOiAc": "Collapse",
   "PRnqYA": "Required. The number of seconds to add. Can be negative to subtract seconds.",
@@ -529,6 +532,7 @@
   "U3guPp": "Group by Connector",
   "U3iWVd": "Generates an array of integers starting from a certain number",
   "U56j9a": "(UTC-06:00) Easter Island",
+  "UAwAAS": "Show inputs",
   "UCNM4L": "To reference a parameter, use the dynamic content list.",
   "UEryJE": "Delete Workflow Action",
   "UHCVNK": "Replaces a string with a given string",
@@ -559,6 +563,7 @@
   "VptXzY": "Use \"{value}\" as a custom value",
   "W070M2": "of {max}",
   "W6FdMh": "Required. The name of the new property.",
+  "WCrlES": "Show outputs",
   "WGwH45": "Clear",
   "WLj8DH": "Please enter a valid dictionary",
   "WOOYW1": "Actions perform operations on data or communicate between systems to execute.",
@@ -637,6 +642,7 @@
   "_/VcZ9g.comment": "Description about hybrid connectors",
   "_/WW7If.comment": "True",
   "_/doURb.comment": "Label for description of custom array Function",
+  "_/n13VL.comment": "Properties text",
   "_/qchXQ.comment": "Required collection parameter to apply join function on",
   "_/qu3zt.comment": "Hour of the day",
   "_/rjP4U.comment": "Timezone value ",
@@ -660,6 +666,7 @@
   "_0hxr5j.comment": "Timezone value ",
   "_0i/6TR.comment": "Required text parameter to upper case",
   "_0meqK1.comment": "Timezone value ",
+  "_0oebOm.comment": "Outputs text",
   "_0p+pJq.comment": "Label for description of custom mod Function",
   "_0qV0Qe.comment": "Required text parameter to apply indexOf function on",
   "_0qcKzG.comment": "Unknown Parameter error message",
@@ -1066,6 +1073,7 @@
   "_PBvR9Q.comment": "Timezone value ",
   "_PF87Ew.comment": "This is a status message to be shown in a monitoring view",
   "_PKMkb/.comment": "Required number parameter to be summed in add function",
+  "_PORNMZ.comment": "Inputs text",
   "_PP63jY.comment": "This is a label to access the overflowed panels",
   "_PQOiAc.comment": "An accessible label for collapse toggle icon",
   "_PRnqYA.comment": "Required integer parameter to subtract seconds from time",
@@ -1146,6 +1154,7 @@
   "_U3guPp.comment": "Label for the checkbox to group results by connector",
   "_U3iWVd.comment": "Label for description of custom range Function",
   "_U56j9a.comment": "Timezone value ",
+  "_UAwAAS.comment": "Show inputs text",
   "_UCNM4L.comment": "Description for Workflow Parameters Part 2",
   "_UEryJE.comment": "Title for operation node",
   "_UHCVNK.comment": "Label for description of custom replace Function",
@@ -1176,6 +1185,7 @@
   "_VptXzY.comment": "Label for button to allow user to create custom value in combobox from current input",
   "_W070M2.comment": "Text on a pager where people can select a page number out of {max}",
   "_W6FdMh.comment": "Required string parameter for new property name in addProperty function",
+  "_WCrlES.comment": "Show outputs text",
   "_WGwH45.comment": "Label to clear editor",
   "_WLj8DH.comment": "Error Message for Invalid Dictionary",
   "_WOOYW1.comment": "Description of what Actions are, on a tooltip about Actions",
