+++ libs/designer-ui/src/lib/monitoring/index.tsx
@@ -1,5 +1,4 @@
 export * from './errorsection';
-export * from './monitoringPanel';
 export * from './requestpanel';
 export * from './retrypanel';
 export * from './statuspill';
