+++ libs/services/intl/src/compiled-lang/strings.en-XA.json
@@ -125,6 +125,12 @@
       "value": "Ḉŏńṿèŕṭ ṭḫè íńṗůṭ ṭŏ âń âŕŕâẏ"
     }
   ],
+  "/n13VL": [
+    {
+      "type": 0,
+      "value": "Ṕŕŏṗèŕṭíèś"
+    }
+  ],
   "/qchXQ": [
     {
       "type": 0,
@@ -261,6 +267,12 @@
     {
       "type": 0,
       "value": "(ŨṮḈ-05:00) Ḃŏĝŏṭâ, Ļíṁâ, Ɋůíṭŏ, Ŕíŏ Ḃŕâńćŏ"
+    }
+  ],
+  "0oebOm": [
+    {
+      "type": 0,
+      "value": "Õůṭṗůṭś"
     }
   ],
   "0p+pJq": [
@@ -2951,6 +2963,12 @@
       "value": "Ŕèɋůíŕèḋ. Ṯḫè ńůṁḃèŕ ṭŏ âḋḋ ṭŏ Ṡůṁṁâńḋ 1."
     }
   ],
+  "PORNMZ": [
+    {
+      "type": 0,
+      "value": "Ḭńṗůṭś"
+    }
+  ],
   "PP63jY": [
     {
       "type": 0,
@@ -3473,6 +3491,12 @@
       "value": "(ŨṮḈ-06:00) Ḕâśṭèŕ Ḭśĺâńḋ"
     }
   ],
+  "UAwAAS": [
+    {
+      "type": 0,
+      "value": "Ṡḫŏẘ íńṗůṭś"
+    }
+  ],
   "UCNM4L": [
     {
       "type": 0,
@@ -3667,6 +3691,12 @@
     {
       "type": 0,
       "value": "Ŕèɋůíŕèḋ. Ṯḫè ńâṁè ŏḟ ṭḫè ńèẘ ṗŕŏṗèŕṭẏ."
+    }
+  ],
+  "WCrlES": [
+    {
+      "type": 0,
+      "value": "Ṡḫŏẘ ŏůṭṗůṭś"
     }
   ],
   "WGwH45": [
