+++ libs/services/intl/src/compiled-lang/strings.json
@@ -125,6 +125,12 @@
       "value": "Convert the input to an array"
     }
   ],
+  "/n13VL": [
+    {
+      "type": 0,
+      "value": "Properties"
+    }
+  ],
   "/qchXQ": [
     {
       "type": 0,
@@ -261,6 +267,12 @@
     {
       "type": 0,
       "value": "(UTC-05:00) Bogota, Lima, Quito, Rio Branco"
+    }
+  ],
+  "0oebOm": [
+    {
+      "type": 0,
+      "value": "Outputs"
     }
   ],
   "0p+pJq": [
@@ -2951,6 +2963,12 @@
       "value": "Required. The number to add to Summand 1."
     }
   ],
+  "PORNMZ": [
+    {
+      "type": 0,
+      "value": "Inputs"
+    }
+  ],
   "PP63jY": [
     {
       "type": 0,
@@ -3473,6 +3491,12 @@
       "value": "(UTC-06:00) Easter Island"
     }
   ],
+  "UAwAAS": [
+    {
+      "type": 0,
+      "value": "Show inputs"
+    }
+  ],
   "UCNM4L": [
     {
       "type": 0,
@@ -3667,6 +3691,12 @@
     {
       "type": 0,
       "value": "Required. The name of the new property."
+    }
+  ],
+  "WCrlES": [
+    {
+      "type": 0,
+      "value": "Show outputs"
     }
   ],
   "WGwH45": [
